Feat: 로그아웃,playlist세팅,playlist추가 페이지와기능 구현
playlist 페이지, 기능 구현.
playlist생성 페이지,기능 구현.
playlist의 리스트조회 예외처리.
유저 로그아웃 기능 구현.
변수cardYoutRegData의 변수형 변경(LocalDateTime).
CardController > PlaylistController 로 클래스명 변경.
home card의 요소(cardYoutRegData)를 포함하도록 수정.
home 네비게이션바 옵션 수정.
playlistUserIdx 변수형 long > int 수정.

+++ src/main/java/com/geonho1943/sharemylist/dto/UserDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public class UserDto {
 
-    private Long userIdx;
+    private int userIdx;
     private String userId;
     private String userName;
     private LocalDateTime userReg;
@@ -15,7 +15,7 @@ public class UserDto {
     // 기본 생성자 및 모든 필드를 포함하는 생성자
     public UserDto() {}
 
-    public UserDto(Long userIdx, String userId, String userName, LocalDateTime userReg, int userPrivileges) {
+    public UserDto(int userIdx, String userId, String userName, LocalDateTime userReg, int userPrivileges) {
         //로그인을 위한 생성자
         this.userIdx = userIdx;
         this.userId = userId;
@@ -25,11 +25,11 @@ public class UserDto {
     }
 
 
-    public Long getUserIdx() {
+    public int getUserIdx() {
         return userIdx;
     }
 
-    public void setUserIdx(Long userIdx) {
+    public void setUserIdx(int userIdx) {
         this.userIdx = userIdx;
     }
